changed the greedy approach for a dp approach

+++ src/main/java/com/mobiquity/factory/ItemFactory.java
@@ -4,5 +4,5 @@ import com.mobiquity.model.ItemDTO;
 
 public class ItemFactory {
 
-    public static ItemDTO createItem(Integer id, Double weight, Double value) {return new ItemDTO(id, weight, value);}
+    public static ItemDTO createItem(Integer id, Double weight, Integer value) {return new ItemDTO(id, weight, value);}
 }
